server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.7.1+cpu",
    "torchvision==0.22.1+cpu",
    "torchaudio==2.7.1+cpu",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "filelock==3.18.0",
    "fsspec==2025.5.1",
    "hf-xet==1.1.3",
    "huggingface-hub==0.32.4",
    "idna==3.10",
    "jinja2==3.1.4",
    "markupsafe==2.1.5",
    "mpmath==1.3.0",
    "networkx==3.3",
    "numpy==2.2.6",
    "packaging==25.0",
    "pillow==11.0.0",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "safetensors==0.5.3",
    "sentencepiece==0.2.0",
    "setuptools==70.2.0",
    "sympy==1.13.3",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "transformers==4.52.4",
    "typing-extensions==4.14.0",
    "urllib3==2.4.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.34.3",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]


[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
torchaudio = [{ index = "pytorch-cpu" }]
